# claimguardian-florida/services/api/openapi.yaml
openapi: 3.1.0
info:
  title: ClaimGuardian Florida API
  version: 1.0.0
  description: Florida property intelligence and claims optimization API
servers:
  - url: https://api.claimguardian.com
    description: Production
  - url: http://localhost:3000/api
    description: Development
paths:
  /parcels/{id}:
    get:
      summary: Get parcel details by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Parcel identifier
      responses:
        "200":
          description: Parcel record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parcel"
        "404":
          description: Parcel not found
...
components:
  schemas:
    Parcel:
      type: object
      properties:
        parcel_id: { type: string }
        county_fips: { type: string }
        address: { type: string }
        owner_name: { type: string }
        just_value: { type: number, format: float }
        assessed_value: { type: number, format: float }
        site_centroid: 
          type: object
          properties:
            lat: { type: number }
            lon: { type: number }
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []